Work with View Controllers (completed)
Start Developing iOS Apps (Swift) -> Building the UI -> Work with View Controllers is now complete. I ran into an issue as this guide is outdated, being built for swift 4 and XCode 9ish  while I'm running Swift 5 & XCode 11. I had to troubleshoot a bit and rewrite code it gave to me that was supposed to work but managed it even though it was unlike anything I've ever worked with before.