Swift: Delegations & Data Sources -- Completed
I completed the course Swift: Delegations & Data Sources, I now generally understand what delegates are and how to manage multiple views. Using the exercise files from that course as a reference, I will try to build simple a iOS storyboard. It will switch tp a shop view when a button is clicked, allows for selection of different items in the shop view, and then passes items back to original screen to a shopping cart. Will expand upon it even further once that is completed.