Fixed issue with delegate not functioning
Copied from a chat I just sent -- I figured out that issue we left off on with the override func prepare(for segue:) not being called, leaving the delegate as nil in the secondary view controller and subsequently crashing when the done button was hit. It seemingly came down to the main storyboard (the one with the open shop view button) in the actual story board view having it's custom class property set as a UIViewController instead of a MainViewController (the custom class),  so when the app was launched and it's view set to the main controller the specified code inside of the MainViewController class was never called due to that view being left as a generic UIViewController. Very similar situation to what happened with the back button not working, a small property value among many different pages of options being set to the wrong value, devil is in the details.