View Controllers
Created view controllers for all of the Dogs section of the tabbed app (i.e. Home, Dogs, Settings at the bottom and you choose a screen). I started to build up code inside of some of the deepest VCs and started to work my way back to link this whole section together. Put off work on building a timer management system as it is complicated and can be done later when system is more built and I am more in the groove.