I DID IT
this was an extremely annoying problem but I figured it out. Typically to pass information back up view controllers you use delegates, complex process but basically since the child view controller does not know of the exsistance of the parent it is how you do it. I had a problem passing a delegate back because a UINavigationController was in the way and it allowed for no conventional, or to me atleast known, way of linking the delegate so the information attempting to be passed can successfully be linked, handled, and accepted. After lots of digging and lots of trial an error, I found how to get the information to pass through the middle man that is the UINavigationController. Way way way to hard to explain in this short of a text but I tried.