progresss
Everything is connected in the dogs tab. All that is left for this tab is to build up the second table view controller so that you can see the different dogs and enable/disable them. Then I'm working on the other tabs, timers (so it actually rings when its time, and persistance so data isnt deleted when the app is closed.