Headache and a half
I made editting the dog work along with a trash button, I thought it would be easy but I needed to do a lot of work passing data through and back and additionally making sure I did it properly otherwise stuff wouldn't update to reflect changes. Pointed out important part of having get and set methods instead of directly accessing the variable, because with get and set methods, everytime the variable is changed you can execute code to reflect the change. This week has been a grind, but I have done so willingly and my product is being expedited and turned out (at around 8-9 hours of work so far this week, along with managing all of my other classes).