Bug hunting
I pushed the app to my phone and found a couple bugs and solved them. I also had a big revelation with my data model, since I use classes it is all reference types and not value types, so if you set a new var equal to one of my old classes, if you change something in the new one it also changes the old one (i.e. emailing someone a text file (value type) vs emailing them a google doc (reference type)). So I knew about this, but didnt account for it in some areas. I could heavily use this to my advantage I believe when changing data, so I don't have to use as many delegates and such but I believe that is more of a 'hacky' way to do it. Since I ran into some issues with all of that, I incorporated NSCopying into my model so I can do "var newDogManager: DogManager = oldDogManager.copy()" and with the .copy() the newDog is given its completely own reference of oldDogManager, any changes to it will NOT have any effect on oldDogManager. If I did that same thing but without the .copy(), any changes to newDogManager would affect oldDogManager the same.