Timing Done (for now)
Timing Class is done for now, lots of compelx interactions involved that pointed out flaws where I slacked in my model and consistency design. Solved most of those to allow for a consistent and more abstract product. Currently you can add a dog and its requirements, edit the dog (cant edit requirements but that easy to add, delete a dog, delete a singular requirement, turn a dog on/off, turn a requirement on/off, pause all alarms, and the app will pop up an inapp notication when one of the alarms goes off. Data is consistent accross the app, it is properly updated when from any point in the app some value is changed and it is extremely close to a MVP. All I want to add now is a schedule pause all alarm function (e.g. between 10pm and 8 am pause all alarms, if an alarm has 5 minutes left at 10 pm when it hits 8 am it will have 5 minutes left) and a proper main menu display that chronologically displays the alarms that will activate the soonest.