caveats fixed
Few caveats fixed with Timing, now when all alarms are paused the time elasped by any active time is saved to the actual requirement, this is reset when the timer is turned on/off or executed. Allows for a timer to properly save its time remaining if the Pause All Alarms button is turned on and off repeatedly in one iteration of an alarm.