Significant UI and timing improvement
Reviewed app with my dad again and made list of improvements. Spent lots of time unifying and improving ui elements along with making naming of functionality more clear and consistant. I have started on functionality difference since that stuff was generally quick. Revamped alertController presentation system, built a system that can queue alertControllers. Reworked timers change how they function, instead of restarting the instant they fire, they wait for a user input. User is prompted to specify an action when a timer fires: disable, snooze, or "did it". This either disables the alarm, snoozes it for a default/Constant time (e.g. I have the default set for 30 minutes) or logs the alarm as done and restarts for the normal interval. Added a nudge for when a dog attempted to be added with no reminders. Next up is adding the ability to edit a requirement. After that the log button on the home screen (different from the three log actions when an alertController is presented for when a reminder's time is up) will be modified to allow for selection of any of those three different option, currently it only has the ability to reset ("Log") a timer.