Requirement Editting
I forgot everything I've done today but so far I've added requirement editing into both area it was needed. Last up on my checklist of things to do before another checkin on the state of the app is to add a way to select one of the three logging options (disable, snooze, or done) to the main screen.