Persistence and notifications
Both are functional and relatively bug free, have been looking into ios control flow and hammering it out to make sure everything is working properly. For now everything seems alright, need to update settings page to add an option to enable / disable notifications within the app as currently there is only a prompt when the app is added for the first time.