Time Of Day Complete
Time of Day is implemented both in backend and UI, it seems to be working to a decent capacity but I have not extensively tested. Added UI animations dogs add page with programmically added buttons, so now when you click the plus buttons it shows mini plus buttons with the options: Add Dog, Add Requirement to "DOG 1", Add Requirement to "DOG 2". Buttons are linked properly and conform / adapt to device size and number of dogs so it won't go off screen and is capped at what I calculate as the max capacity buttons (i.e. will only show as many buttons as space for on the device).