Polish done
Pretty much everything done for now, meeting with my dad tomorrow. There is one small animation situation to fix that is really bizare but I'm holding off for now as I might redo / change the area that is in to a different setup for that functionality. But I added some more settings and made some things more formal, e.g. if there are no reminders created then the pause all reminders switch in the settings is disabled and set to off, etc..