Discovered critical bug while refactoring
While moving the calculations for the executionDate of a requirement from TimingManager to inside of the requirement, where it should be, I discovered a bug that messed up when the timer (aka alarms) fired if toggle pause was ultilized.  Fixed it and did testing to relatively ensure that timing functionality with a variety of inputs is working properly, and completed migration of that calculation. This migration will make it much easier to implement something I was trying to earlier but stopped when I ran into lots of weird spagetti code