API Progress, /login and /dogs heavily revised
Revised initial GET, POST, PUT, DELETE methods to properly interface with database.

Made promise query function to remove nasty nesting of multiple database queries and streamline database query process.

Implemented middleware to validate the format and permissions of the parameters and body. Previous commit was doing it in every controller.

API still lacks proper authentication, but everything is in place for a token system. Once token is in place, all it takes is adding a piece of middleware to all paths.