Hound Endpoint Progress
Utils for actually sending requests to the api for users, dogs, reminders, and logs are nearly finalized. Currently, the main goal is converting everything from dogName, reminderUUID, and logUUID to dogId, reminderId, and logId. This means every single function in the code must be modified to fit this new format. Once that is done, the refreshed code can begin to make actual server calls to sync its data.