Single-User Prototype Near Completion
Need to add Sign In With Apple button
Need to make edit dog/reminder pages pull from server first
Otherwise, all information syncs with the server when its created, updated, or deleted. Error handling is also taken care of so user is alerted properly and actions are reverted if issues are experienced with the queries.