local dogIcons complete
Instead of uploading dogIcons to the server, slowing down queries & making their size MBs instead of Bytes or KBs, there is smart local persistence of dogIcons. Whenever the sever is successfully queried to get, post, put, or delete a dog, the dogIcon on the local Hound device is applied to the retrieved dog, saved to local storage, has its local icon updated, or is deleted. The only downside of this approach is that dogIcons will be lost if a user deletes the app (but family members can share one dog and all have different dogIcons).