Pause All Reminders fixed for family mode
re-enabled pause all reminders for family mode. everything from ispaused, lastpause, lastunpause, and updates to reminders are all synced when the user hits the pause all reminders button.

additionally, restructured API request classes in swift to make it more seamless and clean. makes it easy to select whether or not you want to invoke an error message if the query fails and reduced a lot of the file sizes (~10%-50%) by moving certain functions to one centralized point.