Client-side, server-synced persistance
Client now saves all data it retrieves locally. this means that when it goes to fetch a dogManager or a single dog from the server, it will only recieve back the entries that were changed since it lasted performed a similar query. Therefore, instead of receiving the entire dogs/reminders/logs every single time, the results of past are persisted and only new information is retrieved, reducing network congestion & speeding up queries.