Graceful catch if database server down.
Instead of forcing the node server to crash in the event that any of its databaseConnections cannot reach the database, we instead just gracefully log the event. This stops an infinite loop of force crashing / restarting from happening. Instead, the Hound server be kept alive by PM2 and simply send failure responses to any request. When the database is back online, then Hound will be able to aquire new connections and process requests